
\name{SQLAccess}

\alias{getSwath}
\alias{getFeaturesFlex}
\alias{getMZRTVol}
\alias{getFromMap}
\alias{getMS1id}
\alias{getRTAxis}
\alias{summaryDB}
\alias{mapSummary}
\alias{swathMS1Coverage}
\alias{getXIC}

\title{Basic database access function}


\description{
Database access functions
}

\author{
Witold Wolski
}

\examples{


}

\references{


}
