
\name{SQLPrepare}

\alias{createConnection}
\alias{createBestFeaturesView}
\alias{addIndexMZRT2Features}
\alias{addIndexRT2Features}
\alias{dropIndexFeatures}
\alias{updateSwathWindows}

\title{Database preparation functions}


\description{
Database preparation functions
}

\author{
Witold Wolski
}

\examples{}

\references{


}
